Rename to Crypto.Protocol.DH
